Fix + Improve MySQLSchema sample

+++ mysql-schema/src/main/java/io/javaoperatorsdk/operator/sample/MySQLSchemaController.java
@@ -17,12 +17,14 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
 import java.util.Base64;
+
+import org.apache.commons.lang3.ObjectUtils;
 import org.apache.commons.lang3.RandomStringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 @Controller
-public class SchemaController implements ResourceController<Schema> {
+public class MySQLSchemaController implements ResourceController<MySQLSchema> {
   static final String USERNAME_FORMAT = "%s-user";
   static final String SECRET_FORMAT = "%s-secret";
 
@@ -30,12 +32,12 @@ public class SchemaController implements ResourceController<Schema> {
 
   private final KubernetesClient kubernetesClient;
 
-  public SchemaController(KubernetesClient kubernetesClient) {
+  public MySQLSchemaController(KubernetesClient kubernetesClient) {
     this.kubernetesClient = kubernetesClient;
   }
 
   @Override
-  public UpdateControl<Schema> createOrUpdateResource(Schema schema, Context<Schema> context) {
+  public UpdateControl<MySQLSchema> createOrUpdateResource(MySQLSchema schema, Context<MySQLSchema> context) {
     try (Connection connection = getConnection()) {
       if (!schemaExists(connection, schema.getMetadata().getName())) {
         try (Statement statement = connection.createStatement()) {
@@ -99,7 +101,7 @@ public class SchemaController implements ResourceController<Schema> {
   }
 
   @Override
-  public DeleteControl deleteResource(Schema schema, Context<Schema> context) {
+  public DeleteControl deleteResource(MySQLSchema schema, Context<MySQLSchema> context) {
     log.info("Execution deleteResource for: {}", schema.getMetadata().getName());
 
     try (Connection connection = getConnection()) {
@@ -109,11 +111,13 @@ public class SchemaController implements ResourceController<Schema> {
         }
         log.info("Deleted Schema '{}'", schema.getMetadata().getName());
 
-        if (userExists(connection, schema.getStatus().getUserName())) {
-          try (Statement statement = connection.createStatement()) {
-            statement.execute(format("DROP USER '%1$s'", schema.getStatus().getUserName()));
+        if (schema.getStatus() != null) {
+          if (userExists(connection, schema.getStatus().getUserName())) {
+            try (Statement statement = connection.createStatement()) {
+              statement.execute(format("DROP USER '%1$s'", schema.getStatus().getUserName()));
+            }
+            log.info("Deleted User '{}'", schema.getStatus().getUserName());
           }
-          log.info("Deleted User '{}'", schema.getStatus().getUserName());
         }
 
         this.kubernetesClient
@@ -134,13 +138,20 @@ public class SchemaController implements ResourceController<Schema> {
   }
 
   private Connection getConnection() throws SQLException {
-    return DriverManager.getConnection(
-        format(
-            "jdbc:mysql://%1$s:%2$s?user=%3$s&password=%4$s",
+    if (ObjectUtils.anyNull(System.getenv("MYSQL_HOST"),
+            System.getenv("MYSQL_USER"), System.getenv("MYSQL_PASSWORD"))) {
+      throw new IllegalStateException("Mysql server parameters not defined");
+    }
+
+    String connectionString = format(
+            "jdbc:mysql://%1$s:%2$s",
             System.getenv("MYSQL_HOST"),
-            System.getenv("MYSQL_PORT") != null ? System.getenv("MYSQL_PORT") : "3306",
-            System.getenv("MYSQL_USER"),
-            System.getenv("MYSQL_PASSWORD")));
+            System.getenv("MYSQL_PORT") != null ? System.getenv("MYSQL_PORT") : "3306");
+    String user = System.getenv("MYSQL_USER");
+    String password = System.getenv("MYSQL_PASSWORD");
+
+    log.info("Connecting to '{}' with user '{}'", connectionString, user);
+    return DriverManager.getConnection(connectionString, user, password);
   }
 
   private boolean schemaExists(Connection connection, String schemaName) throws SQLException {
@@ -149,7 +160,7 @@ public class SchemaController implements ResourceController<Schema> {
             "SELECT schema_name FROM information_schema.schemata WHERE schema_name = ?")) {
       ps.setString(1, schemaName);
       try (ResultSet resultSet = ps.executeQuery()) {
-        return resultSet.first();
+        return resultSet.next();
       }
     }
   }
